adk/requirements.txt
# MCP and FastMCP dependencies
mcp>=1.0.0
fastmcp>=0.1.0

# HTTP server dependencies
uvicorn>=0.24.0
fastapi>=0.104.0

# Google ADK dependencies
google-adk>=0.1.0

# HTTP client dependencies
httpx>=0.25.0
aiohttp>=3.9.0

# Logging and utilities
python-dotenv>=1.0.0

# For health checks
curl_cffi>=0.5.0 